Do not block while blinking
This addresses an issue when the no SD card blinks and the ZuluSCSI
V6 Util is open. Before it would cause the Util to freeze while the
LEDs was blinking. Now it does not.

Added the ability to blink without blocking
Requires a polling function in the main loop
Added the ability to cancel blinking
While the LED is blinking disable other blink requests